# BACK/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "안전 경로 추천 AI: ZIP_R0",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://zip-r0.vercel.app"
    }
  ],
  "paths": {
...
    "/emergency/siren/me": {
      "post": {
        "tags": ["Emergency & Disguise"],
        "summary": "Trigger Siren Me",
        "parameters": [
          { "name": "lat", "in": "query", "schema": { "type": "number" } },
          { "name": "lng", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Successful Response" } }
      }
    },
    "/emergency/disguise/me": {
      "post": {
        "tags": ["Emergency & Disguise"],
        "summary": "Trigger Disguise Me",
        "parameters": [
          { "name": "lat", "in": "query", "schema": { "type": "number" } },
          { "name": "lng", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Successful Response" } }
      }
    }
  }
}